requirements.txt
mcp>=1.27.0
starlette>=0.36.3
uvicorn>=0.27.1
google-auth>=2.28.1
google-cloud-securitycenter>=1.33.0
google-cloud-logging>=3.9.0
google-cloud-aiplatform>=1.42.1
google-secops-mcp>=0.7.0
requests>=2.31.0
boto3>=1.34.0
google-cloud-monitoring>=2.20.0
secops>=0.39.0
PyYAML>=6.0
google-cloud-secret-manager>=2.20.0
